WIP: propagating pointers

+++ src/legacy/cxgo_GAME1_1.go
@@ -677,7 +677,7 @@ func nox_xxx_cliPlrInfoLoadFromFile_41A2E0(path *byte, pind int32) int32 {
 	}
 	*memmap.PtrUint16(0x5D4594, 527696) = uint16(nox_xxx_unitGetHP_4EE780(v3))
 	*memmap.PtrUint32(0x5D4594, 527696) = uint32(*memmap.PtrUint16(0x5D4594, 527696))
-	*memmap.PtrUint16(0x5D4594, 527700) = uint16(nox_xxx_unitGetOldMana_4EEC80(v3))
+	*memmap.PtrUint16(0x5D4594, 527700) = nox_xxx_unitGetOldMana_4EEC80(v3)
 	*memmap.PtrUint32(0x5D4594, 527700) = uint32(*memmap.PtrUint16(0x5D4594, 527700))
 	sub_4EFF10(v3)
 	sub_419E10(v3, 1)
